// src/protos/google/api/http.ts
export interface HttpRule {
    /**
     * Selects a method to which this rule applies.
     *
     * Refer to [selector][google.api.DocumentationRule.selector] for syntax details.
     *
     * @generated from protobuf field: string selector = 1;
     */
    selector: string;
    /**
     * @generated from protobuf oneof: pattern
     */
    pattern: {
        oneofKind: "get";
        /**
         * Maps to HTTP GET. Used for listing and getting information about
         * resources.
         *
         * @generated from protobuf field: string get = 2;
         */
        get: string;
    } | {
        oneofKind: "put";
        /**
         * Maps to HTTP PUT. Used for replacing a resource.
         *
         * @generated from protobuf field: string put = 3;
         */
        put: string;
    } | {
        oneofKind: "post";
        /**
         * Maps to HTTP POST. Used for creating a resource or performing an action.
         *
         * @generated from protobuf field: string post = 4;
         */
        post: string;
    } | {
        oneofKind: "delete";
        /**
         * Maps to HTTP DELETE. Used for deleting a resource.
         *
         * @generated from protobuf field: string delete = 5;
         */
        delete: string;
    } | {
        oneofKind: "patch";
        /**
         * Maps to HTTP PATCH. Used for updating a resource.
         *
         * @generated from protobuf field: string patch = 6;
         */
        patch: string;
    } | {
        oneofKind: "custom";
        /**
         * The custom pattern is used for specifying an HTTP method that is not
         * included in the `pattern` field, such as HEAD, or "*" to leave the
         * HTTP method unspecified for this rule. The wild-card rule is useful
         * for services that provide content to Web (HTML) clients.
         *
         * @generated from protobuf field: google.api.CustomHttpPattern custom = 8;
         */
        custom: CustomHttpPattern;
    } | {
        oneofKind: undefined;
    };
    /**
     * The name of the request field whose value is mapped to the HTTP request
     * body, or `*` for mapping all request fields not captured by the path
     * pattern to the HTTP body, or omitted for not having any HTTP request body.
     *
     * NOTE: the referred field must be present at the top-level of the request
     * message type.
     *
     * @generated from protobuf field: string body = 7;
     */
    body: string;
    /**
     * Optional. The name of the response field whose value is mapped to the HTTP
     * response body. When omitted, the entire response message will be used
     * as the HTTP response body.
     *
     * NOTE: The referred field must be present at the top-level of the response
     * message type.
     *
     * @generated from protobuf field: string response_body = 12;
     */
    responseBody: string;
    /**
     * Additional HTTP bindings for the selector. Nested bindings must
     * not contain an `additional_bindings` field themselves (that is,
     * the nesting may only be one level deep).
     *
     * @generated from protobuf field: repeated google.api.HttpRule additional_bindings = 11;
     */
    additionalBindings: HttpRule[];
}
/**
 * A custom pattern is used for defining custom HTTP verb.
 *
 * @generated from protobuf message google.api.CustomHttpPattern
 */
export interface CustomHttpPattern {
    /**
     * The name of this custom HTTP verb.
     *
     * @generated from protobuf field: string kind = 1;
     */
    kind: string;
    /**
     * The path matched by this custom verb.
     *
     * @generated from protobuf field: string path = 2;
     */
    path: string;
}
// ...
class HttpRule$Type extends MessageType<HttpRule> {
    constructor() {
        super("google.api.HttpRule", [
            { no: 1, name: "selector", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "get", kind: "scalar", oneof: "pattern", T: 9 /*ScalarType.STRING*/ },
            { no: 3, name: "put", kind: "scalar", oneof: "pattern", T: 9 /*ScalarType.STRING*/ },
            { no: 4, name: "post", kind: "scalar", oneof: "pattern", T: 9 /*ScalarType.STRING*/ },
            { no: 5, name: "delete", kind: "scalar", oneof: "pattern", T: 9 /*ScalarType.STRING*/ },
            { no: 6, name: "patch", kind: "scalar", oneof: "pattern", T: 9 /*ScalarType.STRING*/ },
            { no: 8, name: "custom", kind: "message", oneof: "pattern", T: () => CustomHttpPattern },
            { no: 7, name: "body", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 12, name: "response_body", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 11, name: "additional_bindings", kind: "message", repeat: 1 /*RepeatType.PACKED*/, T: () => HttpRule }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message google.api.HttpRule
 */
export const HttpRule = new HttpRule$Type();
// @generated message type with reflection information, may provide speed optimized methods
class CustomHttpPattern$Type extends MessageType<CustomHttpPattern> {
    constructor() {
        super("google.api.CustomHttpPattern", [
            { no: 1, name: "kind", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "path", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
}
/**
 * @generated MessageType for protobuf message google.api.CustomHttpPattern
 */
export const CustomHttpPattern = new CustomHttpPattern$Type();